can copy delete

+++ web/core/schema.py
@@ -1,22 +1,27 @@
 import graphene
 from graphene_django import DjangoObjectType
 from django.contrib.auth.models import User
-from core.models import Product
+from core.models import Product, random_id
 from django.db.models import Q
 from django.contrib.auth import authenticate, login, logout
+
+class AuthenticatedUserType(DjangoObjectType):
+    class Meta:
+        model = User
+        fields = ('id', 'username', 'first_name', 'last_name', 'email')
 
 class UserType(DjangoObjectType):
     class Meta:
         model = User
-        fields = ('username', 'first_name', 'last_name', 'email')
+        fields = ('id', 'first_name',)
 
 class ProductType(DjangoObjectType):
     class Meta:
         model = Product
-        fields = ('id', 'name', 'date', 'file', 'user', 'public', 'description')
+        fields = ('id', 'name', 'date', 'file', 'public', 'description', 'owner')
 
 class Query(graphene.ObjectType):
-    user = graphene.Field(UserType)
+    user = graphene.Field(AuthenticatedUserType)
     products = graphene.List(ProductType)
     product = graphene.Field(ProductType, id=graphene.String(required=True))
     def resolve_user(root, info):
@@ -25,7 +30,7 @@ class Query(graphene.ObjectType):
         else: return None
     def resolve_products(root, info):
         if info.context.user.is_authenticated: 
-            return Product.objects.filter(Q(public=True) | Q(user=info.context.user))
+           return Product.objects.filter(Q(public=True) | Q(owner=info.context.user))
         else: return Product.objects.filter(public=True)
     def resolve_product(root, info, id):
         try: return Product.objects.get(id=id)
@@ -39,7 +44,7 @@ class Login(graphene.Mutation):
     @classmethod
     def mutate(cls, root, info, username, password):
         user = authenticate(username=username, password=password)
-        if user is not None: login(info.context, user)
+        if user: login(info.context, user)
         return Login(user=user)
 
 class Logout(graphene.Mutation):
@@ -59,10 +64,15 @@ class CopyProduct(graphene.Mutation):
     product = graphene.Field(ProductType)
     @classmethod
     def mutate(cls, root, info, id, name):
-        product = Product.objects.get(id=id)
-        product.id = None
-        product.name = name
-        product.save()
+        product = None
+        if info.context.user.is_authenticated:  
+            product = Product.objects.get(id=id)
+            if product:  
+                product.id = None
+                product.name = name
+                product.owner = info.context.user
+                product.public = False
+                product.file.save(random_id() + '.glb', product.file, save = True)
         return CopyProduct(product=product)
 
 class DeleteProduct(graphene.Mutation):
@@ -71,8 +81,8 @@ class DeleteProduct(graphene.Mutation):
     product = graphene.Field(ProductType)
     @classmethod
     def mutate(cls, root, info, id):
-        product = Product.objects.get(id=id)
-        product.delete()
+        product = Product.objects.get(id=id, owner=info.context.user)
+        if product: product.delete()
         return DeleteProduct(product=product)
 
 class Mutation(graphene.ObjectType):
@@ -84,6 +94,17 @@ class Mutation(graphene.ObjectType):
 schema = graphene.Schema(query=Query, mutation=Mutation)
 
 
+
+#owner_name = graphene.String()
+    #@staticmethod
+    #def resolve_owner_name(root, info, **kwargs):
+    #    return 'Hello World!'
+
+# @classmethod
+#     def get_queryset(cls, queryset, info):
+#         if info.context.user.is_anonymous:
+#             return queryset.filter(published=True)
+#         return queryset
 
 #login = graphene.Field(UserType, username=graphene.String(required=True), password=graphene.String(required=True))
     #logout = graphene.Field(UserType)
